requirements.txt
asgiref==3.2.3
certifi==2019.11.28
chardet==3.0.4
defusedxml==0.6.0
Django==3.0
django-allauth==0.41.0
django-appconf==1.0.3
django-countries==5.5
django-crispy-forms==1.8.1
django-extensions==2.2.5
django-tempus-dominus==5.1.2.13
gunicorn==20.0.4
idna==2.9
oauthlib==3.1.0
Pillow==7.0.0
pkg-resources==0.0.0
python3-openid==3.1.0
pytz==2019.3
requests==2.23.0
requests-oauthlib==1.3.0
six==1.14.0
sqlparse==0.3.0
urllib3==1.25.8
